--- libraries/google/compute/property/regionurlmap_path_matchers_default_route_action_weighted_backend_services_header_action_response_headers_to_add.rb
@@ -1,48 +1,0 @@
-# frozen_string_literal: false
-
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in README.md and
-#     CONTRIBUTING.md located at the root of this package.
-#
-# ----------------------------------------------------------------------------
-module GoogleInSpec
-  module Compute
-    module Property
-      class RegionUrlMapPathMatchersDefaultRouteActionWeightedBackendServicesHeaderActionResponseHeadersToAdd
-        attr_reader :header_name
-
-        attr_reader :header_value
-
-        attr_reader :replace
-
-        def initialize(args = nil, parent_identifier = nil)
-          return if args.nil?
-          @parent_identifier = parent_identifier
-          @header_name = args['headerName']
-          @header_value = args['headerValue']
-          @replace = args['replace']
-        end
-
-        def to_s
-          "#{@parent_identifier} RegionUrlMapPathMatchersDefaultRouteActionWeightedBackendServicesHeaderActionResponseHeadersToAdd"
-        end
-      end
-
-      class RegionUrlMapPathMatchersDefaultRouteActionWeightedBackendServicesHeaderActionResponseHeadersToAddArray
-        def self.parse(value, parent_identifier)
-          return if value.nil?
-          return RegionUrlMapPathMatchersDefaultRouteActionWeightedBackendServicesHeaderActionResponseHeadersToAdd.new(value, parent_identifier) unless value.is_a?(::Array)
-          value.map { |v| RegionUrlMapPathMatchersDefaultRouteActionWeightedBackendServicesHeaderActionResponseHeadersToAdd.new(v, parent_identifier) }
-        end
-      end
-    end
-  end
-end
