+++ libraries/google/compute/property/regionurlmap_path_matchers_path_rules_route_action_weighted_backend_services_header_action.rb
@@ -13,8 +13,8 @@
 #     CONTRIBUTING.md located at the root of this package.
 #
 # ----------------------------------------------------------------------------
-require 'google/compute/property/regionurlmap_path_matchers_path_rules_route_action_weighted_backend_services_header_action_request_headers_to_add'
-require 'google/compute/property/regionurlmap_path_matchers_path_rules_route_action_weighted_backend_services_header_action_response_headers_to_add'
+require 'google/compute/property/regionurlmap_pmprrawbsharht_add'
+require 'google/compute/property/regionurlmap_pmprrawbsharht_add'
 module GoogleInSpec
   module Compute
     module Property
